Add's author config and data entry for the warning
Allow saving the author of the warning in the event that editorjs is
being used by more than one user.

For example when creating an article for a newspaper, the news item
will be reviewed several times by proofreaders, editors and designers,
in each step each role can add warnings in the body of the document.